refactor(player) : 사용자 입력 클래스 변경
