+++ src/main/java/baseball/Application.java
@@ -7,6 +7,6 @@ public class Application {
         final Scanner scanner = new Scanner(System.in);
         // TODO 구현 진행
 
-        Game.playGame(scanner);
+        Player.playGame(scanner);
     }
 }
