+++ src/main/java/baseball/Player.java
@@ -1,12 +1,23 @@
 package baseball;
 
 
+import java.util.Scanner;
+
+import static baseball.Game.makeGameNumber;
+
 public class Player {
 
+    public static void playGame(Scanner scanner){
+        String gameNumber = makeGameNumber();
+        System.out.println(gameNumber);
 
-    public void inputNumber(){
+        while (true){
+            String inputNumber = scanner.nextLine();
 
+            System.out.println("aaa");
+            break;
+        }
     }
 
-
 }
+
